use StarID for `getGaia` and `getHip` too

+++ src/core/modules/Asterism.cpp
@@ -88,7 +88,7 @@ bool Asterism::read(const QString& record, StarMgr *starMgr)
 				}
 				else
 				{
-					asterism[i]=starMgr->searchGaia(static_cast<int64_t>(HP));
+					asterism[i]=starMgr->searchGaia(HP);
 				}
 				
 				if (!asterism[i])
